(Kitty): disabling blinking cursor
I like blinking cursor but I can't make it work in Kitty + Tmux + Vim. It works in Konsole + Tmux (haven't tried Nvim in this combination). It doesn't work even with default configs for Kitty and Tmux. E.g. when starting Kitty with Tmux already attached, exiting and then starting it again, blinking stops.

It's not a blocker though as I've decided on a setup with different active pane backgrounds in Tmux and transparent bg + inactive window dimming in Nvim which is ~perfect.