(Neovim): make LSP (Fidget) notification background transparent.
Note that this removes the bg so if the terminal window is transparent and Neovim is also transparent then the notification area will also be transparent instead of having a solid bg. BUT, this also means that if there's some code behind the notification then it won't be blended with it, the notification area will create a rectangular area where the code will be hidden.