(Kitty): minimize bottom pane margin
Large margin results in Neovim and Kitty sometimes leaving too much empty space on the bottom because there's the required margin and then there's the fact that Nvim's height must be a multiple of the line height. I guess it affects all full-screen apps, but here it' the most noticable.